+++ CHANGELOG.md
@@ -9,6 +9,18 @@ and this project adheres to [Semantic Versioning].
 
 - /
 
+## [0.0.2] - 2024-02-15
+
+### Added
+
+- News content type config yaml files
+- CopyCleanup process plugin
+
+### Changed
+
+- `news.json` body fields updated with 'dirty' content
+- `news.json` adds tags fields
+
 ## [0.0.1] - 2024-02-07
 
 - initial release
@@ -18,4 +30,6 @@ and this project adheres to [Semantic Versioning].
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[0.0.1]: https://github.com/Author/Repository/releases/tag/v0.0.1
+[unreleased]: https://github.com/jphat/ten_mig/compare/v0.0.2...HEAD
+[0.0.2]: https://github.com/jphat/ten_mig/compare/v0.0.1...v0.0.2
+[0.0.1]: https://github.com/jphat/ten_mig/releases/tag/v0.0.1